inventario-service/requirements.txt
flask
pika
psycopg2-binary
python-dotenv
